backend/requirements.txt
Django<=5.2.1
djangorestframework<=3.14.0
djangorestframework-simplejwt<=5.2.2
drf-yasg<=1.21.10

